test(python): offline unit suite in python-bindings workflow
Run cargo test (params parsers) plus pytest unit tests on every
python-bindings job. Keep NCTL as make python-test-nctl / pytest -m nctl.
